# docs/swagger.yaml
basePath: /
definitions:
  domain.Appointment:
    properties:
      amount:
        type: number
      created_at:
        type: string
      doctor:
        $ref: '#/definitions/domain.Doctor'
      doctor_id:
        type: string
      doctor_notes:
        description: filled after consultation
        type: string
      id:
        type: string
      notes:
        description: patient's reason for visit
        type: string
      patient:
        allOf:
        - $ref: '#/definitions/domain.Patient'
        description: Joined fields — populated when we fetch with joins
      patient_id:
        type: string
      slot:
        $ref: '#/definitions/domain.AppointmentSlot'
      slot_id:
        type: string
      status:
        description: pending, confirmed, completed, cancelled
        type: string
      type:
        description: consultation, follow_up, emergency
        type: string
      updated_at:
        type: string
    type: object
  domain.AppointmentSlot:
    properties:
      created_at:
        type: string
      date:
        type: string
      doctor_id:
        type: string
      end_time:
        type: string
      id:
        type: string
      is_booked:
        type: boolean
      start_time:
        type: string
    type: object
...
  domain.BookAppointmentRequest:
    properties:
      notes:
        type: string
      slot_id:
        type: string
      type:
        enum:
        - consultation
        - follow_up
        - emergency
        type: string
    required:
    - slot_id
    - type
    type: object
...
  domain.Doctor:
    properties:
      bio:
        type: string
      consultation_fee:
        type: number
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_email_verified:
        type: boolean
      is_verified:
        type: boolean
      license_number:
        type: string
      name:
        type: string
      phone:
        type: string
      specialty:
        type: string
      updated_at:
        type: string
      years_experience:
        type: integer
    type: object
...
  domain.DoctorAvailability:
    properties:
      created_at:
        type: string
      day_of_week:
        description: 0=Sunday, 1=Monday...6=Saturday
        type: integer
      doctor_id:
        type: string
      end_time:
        description: '"17:00"'
        type: string
      id:
        type: string
      is_active:
        type: boolean
      slot_duration_minutes:
        type: integer
      start_time:
        description: '"09:00"'
        type: string
      updated_at:
        type: string
    type: object
...
  domain.Patient:
    properties:
      address:
        type: string
      created_at:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      gender:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_email_verified:
        type: boolean
      name:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
...
  domain.SetAvailabilityRequest:
    properties:
      day_of_week:
        maximum: 6
        minimum: 0
        type: integer
      end_time:
        description: '"17:00"'
        type: string
      slot_duration_minutes:
        maximum: 120
        minimum: 15
        type: integer
      start_time:
        description: '"09:00"'
        type: string
    required:
    - day_of_week
    - end_time
    - slot_duration_minutes
    - start_time
    type: object
  domain.UpdateAppointmentStatusRequest:
    properties:
      doctor_notes:
        type: string
      status:
        enum:
        - confirmed
        - completed
        - cancelled
        type: string
    required:
    - status
    type: object
...
host: localhost:8080
info:
  contact:
    email: your-email@gmail.com
    name: Emerald
  description: A production-grade healthcare API for Nigeria — appointment booking,
    medical records, and telemedicine
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: HealthConnect API
  version: "1.0"
paths:
  /api/v1/appointments:
    post:
      consumes:
      - application/json
      description: Allows an authenticated patient to book an appointment slot with
        a doctor.
      parameters:
      - description: Booking data for the appointment.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.BookAppointmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Appointment'
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Book an appointment
      tags:
      - Appointments
  /api/v1/appointments/{id}/status:
    patch:
      consumes:
      - application/json
      description: Allows an authenticated doctor to update the status of an appointment
        and add notes.
      parameters:
      - description: The ID of the appointment to update.
        in: path
        name: id
        required: true
        type: string
      - description: Status update and optional doctor notes.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateAppointmentStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Appointment'
      security:
      - BearerAuth: []
      summary: Update appointment status
      tags:
      - Appointments
  /api/v1/appointments/availability:
    post:
      consumes:
      - application/json
      description: Allows an authenticated doctor to set or update their weekly availability
        schedule.
      parameters:
      - description: Availability data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.SetAvailabilityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.DoctorAvailability'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Set doctor availability
      tags:
      - Appointments
  /api/v1/appointments/availability/{doctor_id}:
    get:
      description: Returns a specific doctor's weekly availability schedule. Accessible
        by any authenticated user.
      parameters:
      - description: The ID of the doctor to retrieve availability for.
        in: path
        name: doctor_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.DoctorAvailability'
            type: array
      security:
      - BearerAuth: []
      summary: Get doctor availability
      tags:
      - Appointments
  /api/v1/appointments/doctor/my:
    get:
      description: Returns all appointments for the currently authenticated doctor.
        This endpoint is similar to `/api/v1/appointments/my` but for doctors.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Appointment'
            type: array
      security:
      - BearerAuth: []
      summary: Get doctor's appointments
      tags:
      - Appointments
  /api/v1/appointments/my:
    get:
      description: Returns all appointments for the currently authenticated patient.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Appointment'
            type: array
      security:
      - BearerAuth: []
      summary: Get patient appointments
      tags:
      - Appointments
  /api/v1/appointments/slots/{doctor_id}:
    get:
      description: Returns all available (unbooked) appointment slots for a specific
        doctor on a given date. Accessible by any authenticated user.
      parameters:
      - description: The ID of the doctor to retrieve slots for.
        in: path
        name: doctor_id
        required: true
        type: string
      - description: The date (YYYY-MM-DD) to check for available slots.
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.AppointmentSlot'
            type: array
      security:
      - BearerAuth: []
      summary: Get available appointment slots
      tags:
      - Appointments
  /api/v1/appointments/slots/{doctor_id}/generate:
    post:
      consumes:
      - application/json
      description: Allows an authenticated doctor to generate bookable appointment
        slots for a specific date based on their availability.
      parameters:
      - description: The ID of the doctor to generate slots for.
        in: path
        name: doctor_id
        required: true
        type: string
      - description: The date (YYYY-MM-DD) to generate slots for.
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.AppointmentSlot'
            type: array
      security:
      - BearerAuth: []
      summary: Generate appointment slots
      tags:
      - Appointments
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and your JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"